Refactor camera to map directory
